+++ src/app/component/visit-summary/prescribed-medication/prescribed-medication.component.ts
@@ -6,8 +6,7 @@ import { Observable } from 'rxjs';
 import { DiagnosisService } from '../../../services/diagnosis.service';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { transition, trigger, style, animate, keyframes } from '@angular/animations';
-import { MatSnackBar } from '@angular/material/snack-bar';
-declare var getEncounterProviderUUID: any, getFromStorage: any, getEncounterUUID: any;
+declare var getEncounterUUID: any;
 
 @Component({
   selector: 'app-prescribed-medication',
@@ -57,7 +56,6 @@ export class PrescribedMedicationComponent implements OnInit {
 
   constructor(private service: EncounterService,
     private diagnosisService: DiagnosisService,
-    private snackbar: MatSnackBar,
     private route: ActivatedRoute) { }
 
   searchPrescription = (text$: Observable<string>) =>
@@ -170,9 +168,7 @@ export class PrescribedMedicationComponent implements OnInit {
     } else {
       insertValue = `${insertValue}`;
     }
-    const providerDetails = getFromStorage('provider');
-    const providerUuid = providerDetails.uuid;
-    if (providerDetails && providerUuid === getEncounterProviderUUID()) {
+    if (this.diagnosisService.isSameDoctor()) {
       this.encounterUuid = getEncounterUUID();
       const json = {
         concept: this.conceptMed,
@@ -186,15 +182,17 @@ export class PrescribedMedicationComponent implements OnInit {
           this.meds.push({ uuid: response.uuid, value: insertValue });
           this.add = false;
         });
-    } else { this.snackbar.open('Another doctor is viewing this case', null, { duration: 4000 }); }
+    }
   }
 
   delete(i) {
+    if (this.diagnosisService.isSameDoctor()) {
     const uuid = this.meds[i].uuid;
     this.diagnosisService.deleteObs(uuid)
       .subscribe(res => {
         this.meds.splice(i, 1);
       });
+    }
   }
 
 }
