+++ src/app/component/visit-summary/prescribed-test/prescribed-test.component.ts
@@ -6,8 +6,7 @@ import { DiagnosisService } from 'src/app/services/diagnosis.service';
 import { Observable } from 'rxjs';
 import { debounceTime, distinctUntilChanged, map } from 'rxjs/operators';
 import { transition, trigger, style, animate, keyframes } from '@angular/animations';
-import { MatSnackBar } from '@angular/material/snack-bar';
-declare var getEncounterProviderUUID: any, getFromStorage: any, getEncounterUUID: any;
+declare var getEncounterUUID: any;
 
 @Component({
   selector: 'app-prescribed-test',
@@ -43,7 +42,6 @@ export class PrescribedTestComponent implements OnInit {
 
   constructor(private service: EncounterService,
     private diagnosisService: DiagnosisService,
-    private snackbar: MatSnackBar,
     private route: ActivatedRoute) { }
 
 
@@ -80,9 +78,7 @@ export class PrescribedTestComponent implements OnInit {
     const date = new Date();
     const form = this.testForm.value;
     const value = form.test;
-    const providerDetails = getFromStorage('provider');
-    const providerUuid = providerDetails.uuid;
-    if (providerDetails && providerUuid === getEncounterProviderUUID()) {
+    if (this.diagnosisService.isSameDoctor()) {
       this.encounterUuid = getEncounterUUID();
       const json = {
         concept: this.conceptTest,
@@ -95,15 +91,17 @@ export class PrescribedTestComponent implements OnInit {
         .subscribe(resp => {
           this.tests.push({ uuid: resp.uuid, value: value });
         });
-    } else { this.snackbar.open('Another doctor is viewing this case', null, { duration: 4000 }); }
+    }
   }
 
   delete(i) {
+    if (this.diagnosisService.isSameDoctor()) {
     const uuid = this.tests[i].uuid;
     this.diagnosisService.deleteObs(uuid)
       .subscribe(res => {
         this.tests.splice(i, 1);
       });
+    }
   }
 }
